requirements.txt
aniso8601==9.0.1
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
Flask-RESTful==0.3.9
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
python-telegram-bot==20.1
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
